requirements.txt
python_version==3.8.*
xarray==2023.1.0
netCDF4==1.6.2
cftime==1.6.2
pandas==1.5.3
PyYAML==6.0.1
matplotlib==3.7.2

